refactor && update analytics scripts
