+++ src/components/common/Buttons/SocialButtons/SocialButton.tsx
@@ -14,7 +14,7 @@ type TSocialButton = {
 const SocialButton: React.FC<TSocialButton> = ({icon: IconComponent, tooltipLabel, type}) => {
 	const {hovered, ref} = useHover();
 	const clipboard = useClipboard({timeout: 5000});
-  
+
 	const [url, setUrl] = React.useState(getShareLink(type));
 	const {start: clearClipboard} = useTimeout(() => {
 		clipboard.reset();
